🚚 refactor: toast/tooltip을 components에서 shared/ui로 이동
[변경 내용]
- components/toast(Toast, ToastProvider, useToastStore), components/tooltip(Tooltip)을
  shared/ui/toast, shared/ui/tooltip으로 이동
- 소비자 41곳의 import 경로 갱신(@/components/toast -> @/shared/ui/toast 등)

[변경 이유]
- toast/tooltip은 도메인을 모르는 순수 범용 UI(알림/툴팁)인데 components에 있어
  features -> components 역방향 위반을 26건 유발(useToastStore 35곳 등)
- components 레이어는 features보다 상위라 feature가 toast를 쓰려면 규칙 위반.
  범용 UI의 올바른 위치는 최하층 shared/ui
- boundaries 위반 77 -> 51건 (-26): features -> components 26 -> 1

[고려 사항]
- toast/tooltip은 상위/타 레이어 의존이 전혀 없어(shared/external만 참조) 순수
  이동으로 안전하게 해결. 의존성 역전 불필요
- 내부 상대 import(ToastProvider -> ./Toast, ./useToastStore)는 디렉토리째
  이동이라 그대로 유효
- 전체 파일 sed 치환이라 vi.mock의 문자열 경로(matching 테스트의
  "@/components/tooltip/Tooltip")도 함께 갱신됨
- 남은 component 대상 위반 3건(header<->sidebar 2, intro->header 1)은 도메인을
  아는 조립 컴포넌트라 widgets 레이어 정리 단계에서 처리
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
