+++ src/features/auth/hooks/useEmailVerification.ts
@@ -1,6 +1,5 @@
 import { useEffect, useRef, useState } from "react"
 
-import { useToastStore } from "@/components/toast/useToastStore"
 import {
   completeEmailVerification,
   getEmailAvailability,
@@ -8,6 +7,7 @@ import {
   sendEmailVerification,
 } from "@/features/auth/api/emailVerification"
 import { emailSchema } from "@/features/signup/validation"
+import { useToastStore } from "@/shared/ui/toast/useToastStore"
 
 import type { EmailVerificationPurpose } from "@/features/auth/model/types"
