+++ src/features/settings/ui/AccountSettingsPage.tsx
@@ -2,7 +2,6 @@ import { useQueryClient } from "@tanstack/react-query"
 import { useNavigate } from "@tanstack/react-router"
 import { useRef, useState } from "react"
 
-import { useToastStore } from "@/components/toast/useToastStore"
 import { deleteMember, updateMemberInfo } from "@/entities/member/api/me"
 import { authKeys, useMe } from "@/entities/member/hooks/useMe"
 import { useAuthStore } from "@/entities/member/store/authStore"
@@ -21,6 +20,7 @@ import { Button } from "@/shared/ui/Button"
 import { RoleTagChip } from "@/shared/ui/chip/RoleTagChip"
 import { InputBox } from "@/shared/ui/input/InputBox"
 import { CtaModal } from "@/shared/ui/modal/CtaModal"
+import { useToastStore } from "@/shared/ui/toast/useToastStore"
 
 import { ChangeEmailForm } from "./ChangeEmailForm"
 import { ChangePasswordForm } from "./ChangePasswordForm"
