+++ src/routes/matching/projects/management.tsx
@@ -1,10 +1,10 @@
 import { createFileRoute, redirect, useNavigate } from "@tanstack/react-router"
 import { useEffect } from "react"
 
-import { useToastStore } from "@/components/toast/useToastStore"
 import { canManageProjects } from "@/entities/member/model/identity"
 import { ensureMe } from "@/features/auth/lib/ensureMe"
 import { ProjectManagementPage } from "@/features/project/management"
+import { useToastStore } from "@/shared/ui/toast/useToastStore"
 
 export const Route = createFileRoute("/matching/projects/management")({
   validateSearch: (
